+++ pygtftk/bwig/bw_coverage.py
@@ -11,6 +11,7 @@ from builtins import range
 from builtins import str
 from builtins import zip
 from itertools import repeat
+from tempfile import NamedTemporaryFile
 
 import numpy as np
 from pybedtools import BedTool
@@ -23,6 +24,63 @@ from pygtftk.utils import flatten_list
 from pygtftk.utils import intervals
 from pygtftk.utils import make_tmp_file
 from pygtftk.utils import message
+
+# -------------------------------------------------------------------------
+# TMP_FILE_POOL_MANAGER stores temporary file name
+# make_tmp_file_pool is function that add temporary files to TMP_FILE_POOL_MANAGER
+# TMP_FILE_POOL_MANAGER will be updated by workers (in contrast to a global
+# variable)
+# -------------------------------------------------------------------------
+
+TMP_FILE_POOL_MANAGER = multiprocessing.Manager().list()
+
+
+def make_tmp_file_pool(prefix='tmp',
+                       suffix='',
+                       store=True,
+                       dir=None):
+    """
+    This
+
+    :Example:
+
+    >>> from pygtftk.utils import make_tmp_file_pool
+    >>> tmp_file = make_tmp_file_pool()
+    >>> assert os.path.exists(tmp_file.name)
+    >>> tmp_file = make_tmp_file_pool(prefix="pref")
+    >>> assert os.path.exists(tmp_file.name)
+    >>> tmp_file = make_tmp_file_pool(suffix="suf")
+    >>> assert os.path.exists(tmp_file.name)
+
+    """
+
+    dir_target = None
+
+    if dir is None:
+        if pygtftk.utils.TMP_DIR is not None:
+            if not os.path.exists(pygtftk.utils.TMP_DIR):
+                msg = "Creating directory {d}."
+                message(msg.format(d=pygtftk.utils.TMP_DIR), type="INFO")
+                os.mkdir(pygtftk.utils.TMP_DIR)
+                dir_target = pygtftk.utils.TMP_DIR
+
+    else:
+        if not os.path.exists(dir):
+            msg = "Creating directory {d}."
+            message(msg.format(d=dir), type="INFO")
+            os.mkdir(dir)
+            dir_target = dir
+
+    tmp_file = NamedTemporaryFile(delete=False,
+                                  mode='w',
+                                  prefix=prefix + "_pygtftk_",
+                                  suffix=suffix,
+                                  dir=dir_target)
+
+    if store:
+        TMP_FILE_POOL_MANAGER.append(tmp_file.name)
+
+    return tmp_file
 
 
 # -------------------------------------------------------------------------
@@ -74,7 +132,7 @@ def _big_wig_coverage_worker(input_values):
     else:
         if bin_nb < 1:
             bin_nb = 1
-        matrix_file = make_tmp_file(prefix="worker_coverage_")
+        matrix_file = make_tmp_file_pool(prefix="worker_coverage_")
 
     for cpt, big_wig in enumerate(bw_list):
 
@@ -105,10 +163,8 @@ def _big_wig_coverage_worker(input_values):
             nb += 1
 
             if nb == nb_to_do:
-                message(str(multiprocessing.current_process().name) +
-                        " has processed " +
-                        str(nb) +
-                        " regions")
+                p_name = str(multiprocessing.current_process().name)
+                message(p_name + " has processed " + str(nb) + " regions")
 
             if (i.end - i.start) < bin_nb: